voice_filter/pyproject.toml
[project]
name = "voice-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "audio-separator>=0.39.1",
    "ffmpeg-python>=0.2.0",
    "loguru>=0.7.3",
    "onnxruntime>=1.23.2",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
